Move GeneratePassword to pkg/crypto
